refactor: rename ResourceSavingScene to ResourceSavingView
